# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API of go-ispconfig, the Go port of ISPConfig3. Authenticate via POST /api/login: browsers receive an HTTP-only session cookie (mutating requests then require the X-CSRF-Token header returned by login); non-browser clients present the same session id as a bearer token.",
        "title": "go-ispconfig API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/login": {
            "post": {
                "description": "Verifies sys_user credentials (bcrypt or legacy ISPConfig3 crypt hashes) with brute-force lockout, creates a sys_session and returns the CSRF token and session id. The session id is also set as an HTTP-only cookie; non-browser clients may present it as a bearer token instead. With stay_logged_in the session lives 30 days (sliding) instead of the 1-hour idle TTL.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Log in",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.LoginResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many failed attempts",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes the current sys_session and clears the session cookie. Succeeds with 204 even without a valid session.",
                "tags": [
                    "auth"
                ],
                "summary": "Log out",
                "responses": {
                    "204": {
                        "description": "Logged out"
                    }
                }
            }
        },
...
        "/session": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the logged-in user: username, access level, effective groups, language and the session CSRF token (so a reloaded SPA can resume without logging in again).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Current session",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SessionInfo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/system/scheduler": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists the daemon scheduler jobs with their last-run time and status, read from the sys_config \"scheduler\" mirror the daemon persists after every job execution. Admin only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Scheduler job status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.SchedulerJob"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Not an admin",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ErrorInfo": {
            "type": "object",
            "properties": {
                "fields": {
                    "description": "Fields maps field names to their i18n validation error keys\n(422 responses only).",
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "key": {
                    "description": "Key is the i18n message key describing the error.",
                    "type": "string",
                    "example": "error.permission_denied"
                }
            }
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error describes what went wrong.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.ErrorInfo"
                        }
                    ]
                }
            }
        },
...
        "api.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "description": "Password is the cleartext password.",
                    "type": "string",
                    "example": "secret"
                },
                "stay_logged_in": {
                    "description": "StayLoggedIn requests a long-lived session (30 days sliding\nexpiration, sys_session.permanent) instead of the 1-hour idle TTL.",
                    "type": "boolean"
                },
                "username": {
                    "description": "Username is the sys_user login name.",
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "api.LoginResponse": {
            "type": "object",
            "properties": {
                "csrf_token": {
                    "description": "CSRFToken must be sent as X-CSRF-Token on every mutating\ncookie-authenticated request.",
                    "type": "string"
                },
                "session_id": {
                    "description": "SessionID is the session id, also set as the HTTP-only cookie.\nNon-browser clients present it as \"Authorization: Bearer \u003cid\u003e\".",
                    "type": "string"
                },
                "typ": {
                    "description": "Typ is the access level (\"admin\" or \"user\").",
                    "type": "string",
                    "example": "admin"
                },
                "username": {
                    "description": "Username is the authenticated login name.",
                    "type": "string",
                    "example": "admin"
                }
            }
        },
...
        "api.SchedulerJob": {
            "type": "object",
            "properties": {
                "last_run": {
                    "description": "LastRun is the RFC3339 time of the last execution, empty if never run.",
                    "type": "string",
                    "example": "2026-08-01T03:00:00Z"
                },
                "name": {
                    "description": "Name is the job identifier (\"datalog_prune\").",
                    "type": "string",
                    "example": "datalog_prune"
                },
                "status": {
                    "description": "Status is \"ok\" or \"error: \u003cmessage\u003e\" from the last execution.",
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "api.SessionInfo": {
            "type": "object",
            "properties": {
                "csrf_token": {
                    "description": "CSRFToken is the per-session token for the X-CSRF-Token header, so a\nreloaded SPA can recover it without logging in again.",
                    "type": "string"
                },
                "groups": {
                    "description": "Groups is the effective sys_group id list of the user.",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "language": {
                    "description": "Language is the panel language of the user.",
                    "type": "string",
                    "example": "en"
                },
                "typ": {
                    "description": "Typ is the access level (\"admin\" or \"user\").",
                    "type": "string",
                    "example": "admin"
                },
                "username": {
                    "description": "Username is the sys_user login name.",
                    "type": "string",
                    "example": "admin"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Session id as bearer token: \"Bearer \u003csession id\u003e\". No CSRF token required on this transport.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "description": "Browser transport: goisp_session HTTP-only cookie set by POST /api/login. Mutating requests must also send the X-CSRF-Token header.",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}